infra: float marker timestamps, finish-detect fix, clip window math, extended probe
S0.1  probe_metadata returns 11 new fields (pix_fmt, color_*, audio_*, bitrates)
      so downstream filters can condition on actual source format.
S1.1  marker timestamps store as REAL (sub-second precision). F8 at 12.7s was
      truncating to 12 and shifting the clip window by up to a second.
S1.2  delete dead `markers_temp` flush — early-return in create_marker made
      the queue unreachable, and the flush would have lost F8/F9 kind anyway.
S1.3  marker_worker heartbeat: log press count every 60s so a silent pynput
      death (macOS Accessibility revoked, venv path moved) is visible.
S2.1  remove 10 MB never-finishes gate; replace with 5s start-grace.
S2.2  stable_count 60 -> 20 (log message claimed 10s, code waited 30s).
S2.3  strip Windows-cp1252-crashing emojis from runtime print() strings.
S3.1  anchor clip window end off clamped start. CLIP_DURATION 30 -> 60s for
      TikTok long-form monetization (still <= YT Shorts ceiling).
S3.2  document ClipJob.metadata write-once contract (Router writer; processors
      read-only).

+++ zerino/db/init_db.py
@@ -33,12 +33,16 @@ def create_database():
     # MARKERS
     # `kind` distinguishes hotkey source: F8 = talking_head, F9 = gameplay.
     # Downstream picks the render layout from this (square fill vs face+game split).
+    # `timestamp` is REAL (seconds since recording start) for sub-second
+    # precision — F8 at 12.7s of recording must round-trip as 12.7, not 12,
+    # or the clip window misses the streamer's actual reaction moment by up
+    # to a second. See S1.1 in CLIPPING_QUALITY_PLAN.md.
     cursor.execute("""
     CREATE TABLE IF NOT EXISTS markers (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         streamer_id INTEGER,
         recording_id INTEGER NOT NULL,
-        timestamp INTEGER NOT NULL,
+        timestamp REAL NOT NULL,
         kind TEXT NOT NULL DEFAULT 'talking_head'
             CHECK(kind IN ('talking_head','gameplay')),
         note TEXT,
@@ -49,14 +53,18 @@ def create_database():
     """)
   
     # CLIPS (PIPELINE CORE)
+    # `clip_start` / `clip_end` are REAL seconds so the clip window matches
+    # the float marker.timestamp exactly. ClipService derives them as
+    # `max(0, marker_time - PRE_BUFFER)` and `marker_time + (DURATION-PRE)`;
+    # truncating those to INTEGER lost sub-second alignment with the marker.
     cursor.execute("""
     CREATE TABLE IF NOT EXISTS clips (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         recording_id INTEGER NOT NULL,
         marker_id INTEGER NOT NULL,
         video_file TEXT,
-        clip_start INTEGER NOT NULL,
-        clip_end INTEGER NOT NULL,
+        clip_start REAL NOT NULL,
+        clip_end REAL NOT NULL,
         status TEXT NOT NULL DEFAULT 'pending' CHECK(status IN ('pending','processing','completed','failed')),
         output_path TEXT,
         error_message TEXT,
